+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project (attempts to) adhere to [Semantic Versioning](http://semver.org/).
 
-## [1.2.0] - 2017-19-11
+## [1.2.1] - 2017-10-03
+### Changed
+- Logging now logs mummer commands run with ANIn option
+
+### Added
+- Choose can now function without checkM
+
+## [1.2.0] - 2017-09-11
 ### fixed
 - Scoring with regards to strain heterogeneity modified